create cashier is done
