+++ src/features/cashier-management-dashboard/components/add-cashier/CashierBankDetails.tsx
@@ -1,293 +1,110 @@
-import { ChangeEvent, useState } from 'react';
-import { IoCloudUploadOutline } from 'react-icons/io5';
-import { Link } from 'react-router-dom';
+import { useEffect } from 'react';
+import { toast } from 'react-toastify';
 import { useCashierContext } from '../../layouts/AddCashier';
-import useCashierCRUDService from '../../services/cashierCRUDService';
+import useBankCRUDService from '../../services/BankDetailsCRUDService';
 
-const CashierDetails = () => {
-  const [previewImage, setPreviewImage] = useState<string | null>(null);
+const CashierBankDetails = () => {
+  const { cashierDetails, cashierBankDetails, setCashierBankDetails } =
+    useCashierContext();
 
-  const { cashierDetails, setCashierDetails } = useCashierContext();
+  const { updateBankDetails, loading } = useBankCRUDService();
 
-  const handleImageChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const file: File | null = e.target.files ? e.target.files[0] : null;
-    if (file) {
-      const reader = new FileReader();
-      reader.onloadend = () => {
-        if (typeof reader.result === 'string') {
-          setPreviewImage(reader.result);
-        }
-      };
-      reader.readAsDataURL(file);
-    }
-  };
-  const { createCashier, loading } = useCashierCRUDService();
-
-  const goToBankDetails = () => {
-    // console.log(cashierDetails);
-    createCashier(cashierDetails).then((res) => {
-      if (res) {
-        setCashierDetails({
-          ...cashierDetails,
-          employerId: res,
-        });
-      }
+  useEffect(() => {
+    setCashierBankDetails({
+      ...cashierBankDetails,
+      employerId: cashierDetails.employerId,
+      monthlyPayment: cashierDetails.employerSalary,
     });
+  }, []);
+
+  const goToSummary = () => {
+    console.log('Summary', cashierDetails);
+    if (cashierDetails && cashierDetails.employerId) {
+      updateBankDetails(cashierBankDetails, cashierDetails.employerId);
+    } else {
+      toast.error('No cashier created yet.');
+    }
   };
 
   return (
     <div className='w-full p-16 px-4 sm:px-6 lg:px-8'>
-      <p className='text-2xl font-bold text-center mb-4'>
-        Creating A New Cashier
-      </p>
-      <div className='grid grid-cols-1 md:grid-cols-4 gap-6 items-center justify-center'>
-        <div className='flex items-center justify-center gap-4 flex-col'>
-          {previewImage ? (
-            <div className='mt-4'>
-              <img
-                src={previewImage}
-                alt='Preview'
-                className='w-64 h-64 rounded-full'
-              />
-            </div>
-          ) : (
-            <div className='mt-4'>
-              <img
-                src='https://randomuser.me/api/portraits/men/1.jpg'
-                alt='Preview'
-                className='w-64 h-64 rounded-full'
-              />
-            </div>
-          )}
-          <label className='w-64 flex flex-row items-center p-2 justify-center gap-2 bg-white rounded-lg'>
-            <IoCloudUploadOutline size={25} />
-            <span className='text-base leading-normal'>Select an image</span>
-            <input
-              type='file'
-              className='hidden'
-              onChange={handleImageChange}
-            />
-          </label>
-        </div>
+      <div className='grid grid-cols-1 md:grid-cols-1 gap-6'>
         {/* First Column */}
         <div>
           <label
-            htmlFor='nickname'
+            htmlFor='bankName'
             className='block text-sm font-medium text-black mt-4'
           >
-            Nickname
+            Bank Name
           </label>
           <input
             type='text'
-            id='nickname'
-            className='mt-1 p-2 border-gray rounded-md w-64'
-            value={cashierDetails.employerNicName}
+            id='bankName'
+            className='mt-1 p-2 border-gray rounded-md w-full'
+            value={cashierBankDetails.bankName}
             onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                employerNicName: e.target.value,
+              setCashierBankDetails({
+                ...cashierBankDetails,
+                bankName: e.target.value,
               })
             }
           />
 
           <label
-            htmlFor='nicNumber'
+            htmlFor='branchName'
             className='block text-sm font-medium text-black mt-4'
           >
-            NIC Number
+            Branch Name
           </label>
           <input
             type='text'
-            id='nicNumber'
-            className='mt-1 p-2 border-gray rounded-md w-64'
-            value={cashierDetails.employerNic}
+            id='branchName'
+            className='mt-1 p-2 border-gray rounded-md w-full'
+            value={cashierBankDetails.bankBranchName}
             onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                employerNic: e.target.value,
+              setCashierBankDetails({
+                ...cashierBankDetails,
+                bankBranchName: e.target.value,
               })
             }
           />
 
           <label
-            htmlFor='telephone'
+            htmlFor='accountNumber'
             className='block text-sm font-medium text-black mt-4'
           >
-            Telephone Number
+            Account Number
           </label>
           <input
-            type='tel'
-            id='telephone'
-            className='mt-1 p-2 border-gray rounded-md w-64'
-            value={cashierDetails.employerPhone}
+            type='number'
+            id='accountNumber'
+            className='mt-1 p-2 border-gray rounded-md w-full'
+            value={cashierBankDetails.bankAccountNumber}
             onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                employerPhone: e.target.value,
+              setCashierBankDetails({
+                ...cashierBankDetails,
+                bankAccountNumber: parseInt(e.target.value),
               })
             }
           />
 
           <label
-            htmlFor='email'
+            htmlFor='additionalNotes'
             className='block text-sm font-medium text-black mt-4'
           >
-            Email
+            Additional Notes
           </label>
-          <input
-            type='email'
-            id='email'
-            className='mt-1 p-2 border-gray rounded-md w-64'
-            value={cashierDetails.employerEmail}
+          <textarea
+            id='additionalNotes'
+            className='mt-1 p-2 border-gray rounded-md w-full'
+            value={cashierBankDetails.employerDescription}
             onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                employerEmail: e.target.value,
+              setCashierBankDetails({
+                ...cashierBankDetails,
+                employerDescription: e.target.value,
               })
             }
-          />
-
-          <label
-            htmlFor='firstName'
-            className='block text-sm font-medium text-black mt-4'
-          >
-            First Name
-          </label>
-          <input
-            type='text'
-            id='firstName'
-            className='mt-1 p-2 border-gray rounded-md w-64'
-            value={cashierDetails.employerFirstName}
-            onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                employerFirstName: e.target.value,
-              })
-            }
-          />
-
-          <label
-            htmlFor='lastName'
-            className='block text-sm font-medium text-black mt-4'
-          >
-            Last Name
-          </label>
-          <input
-            type='text'
-            id='lastName'
-            className='mt-1 p-2 border-gray rounded-md w-64'
-            value={cashierDetails.employerLastName}
-            onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                employerLastName: e.target.value,
-              })
-            }
-          />
-        </div>
-
-        {/* Second Column */}
-        <div>
-          <label
-            htmlFor='branch'
-            className='block text-sm font-medium text-black'
-          >
-            Branch
-          </label>
-          <select
-            id='branch'
-            className='mt-1 p-2 border-gray rounded-md w-64'
-            value={cashierDetails.branchId}
-            onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                branchId: parseInt(e.target.value),
-              })
-            }
-          >
-            <option value='0'>Branch 1</option>
-            <option value='1'>Branch 2</option>
-            <option value='2'>Branch 3</option>
-          </select>
-
-          <label
-            htmlFor='gender'
-            className='block text-sm font-medium text-black'
-          >
-            Gender
-          </label>
-          <select
-            id='gender'
-            className='mt-1 p-2 border-gray rounded-md w-64'
-            value={cashierDetails.gender}
-            onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                gender: e.target.value,
-              })
-            }
-          >
-            <option value='MALE'>Male</option>
-            <option value='FEMALE'>Female</option>
-            <option value='OTHER'>Other</option>
-          </select>
-
-          <label
-            htmlFor='addressLine'
-            className='block text-sm font-medium text-black mt-4'
-          >
-            Address
-          </label>
-          <input
-            type='text'
-            id='addressLine1'
-            className='mt-1 p-2 border-gray rounded-md w-64'
-            value={cashierDetails.employerAddress}
-            onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                employerAddress: e.target.value,
-              })
-            }
-          />
-
-          <label
-            htmlFor='dateOfBirth'
-            className='block text-sm font-medium text-black mt-4'
-          >
-            Date of Birth
-          </label>
-          <input
-            type='date'
-            id='dateOfBirth'
-            className='mt-1 p-2 border-gray rounded-md w-64'
-            value={cashierDetails.dateOfBirth.toISOString().split('T')[0]}
-            onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                dateOfBirth: new Date(e.target.value),
-              })
-            }
-          />
-
-          <label
-            htmlFor='role'
-            className='block text-sm font-medium text-black mt-4'
-          >
-            Role
-          </label>
-          <input
-            type='text'
-            id='role'
-            className='mt-1 p-2 border-gray rounded-md w-64'
-            value={cashierDetails.role}
-            onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                role: e.target.value,
-              })
-            }
-          />
+          ></textarea>
 
           <label
             htmlFor='baseSalary'
@@ -298,77 +115,13 @@ const CashierDetails = () => {
           <input
             type='text'
             id='baseSalary'
-            className='mt-1 p-2 border-gray rounded-md w-64'
+            className='mt-1 p-2 border-gray rounded-md w-full'
             value={cashierDetails.employerSalary}
-            onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                employerSalary: parseFloat(e.target.value),
-              })
-            }
-          />
-        </div>
-        {/* Third Column */}
-        <div>
-          <label
-            htmlFor='password'
-            className='block text-sm font-medium text-black mt-4'
-          >
-            Password
-          </label>
-          <input
-            type='text'
-            id='password'
-            className='mt-1 p-2 border-gray rounded-md w-64'
-            value={cashierDetails.employerPassword}
-            onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                employerPassword: e.target.value,
-              })
-            }
-          />
-
-          <label
-            htmlFor='confirmPassword'
-            className='block text-sm font-medium text-black mt-4'
-          >
-            Confirm Password
-          </label>
-          <input
-            type='text'
-            id='confirmPassword'
-            className='mt-1 p-2 border-gray rounded-md w-64'
-            value={cashierDetails.employerConfirmPassword}
-            onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                employerConfirmPassword: e.target.value,
-              })
-            }
-          />
-
-          <label
-            htmlFor='pin'
-            className='block text-sm font-medium text-black mt-4'
-          >
-            Pin
-          </label>
-          <input
-            type='text'
-            id='pin'
-            className='mt-1 p-2 border-gray rounded-md w-64'
-            value={cashierDetails.pin}
-            onChange={(e) =>
-              setCashierDetails({
-                ...cashierDetails,
-                pin: parseInt(e.target.value),
-              })
-            }
+            readOnly
           />
         </div>
       </div>
-      <div className='flex items-center justify-center gap-8 w-full mt-8'>
+      <div className='flex items-center justify-center gap-8 w-full'>
         <button
           type='button'
           className={`text-white py-2.5 px-5 me-2 mb-2 rounded-lg ${
@@ -376,20 +129,14 @@ const CashierDetails = () => {
               ? 'bg-gray-500 cursor-not-allowed'
               : 'bg-blueDarker hover:bg-blue'
           }`}
-          onClick={goToBankDetails}
+          onClick={goToSummary}
           disabled={loading}
         >
-          {loading ? 'Loading...' : 'Create'}
-        </button>
-
-        <button
-          type='button'
-          className='py-2.5 px-5 me-2 mb-2 text-sm font-medium text-slate-900 focus:outline-none bg-white rounded-lg border border-gray hover:bg-gray'
-        >
-          <Link to='/manager-dashboard/Cashiers'>Back To Cashier Manager</Link>
+          {loading ? 'Loading...' : 'Update'}
         </button>
       </div>
     </div>
   );
 };
-export default CashierDetails;
+
+export default CashierBankDetails;
