+++ src/features/cashier-management-dashboard/components/add-cashier/CashierDetailsSummary.tsx
@@ -1,142 +1,92 @@
-import { useEffect } from 'react';
-import { toast } from 'react-toastify';
-import { useCashierContext } from '../../layouts/AddCashier';
-import useBankCRUDService from '../../services/BankDetailsCRUDService';
+import React from 'react';
+import { Link } from 'react-router-dom';
+import { ComponentState, useCashierContext } from '../../layouts/AddCashier';
 
-const CashierBankDetails = () => {
-  const { cashierDetails, cashierBankDetails, setCashierBankDetails } =
+function CashierDetailsSummary() {
+  const { cashierDetails, setCurrentComponent, cashierBankDetails } =
     useCashierContext();
 
-  const { updateBankDetails, loading } = useBankCRUDService();
-
-  useEffect(() => {
-    setCashierBankDetails({
-      ...cashierBankDetails,
-      employerId: cashierDetails.employerId,
-      monthlyPayment: cashierDetails.employerSalary,
-    });
-  }, []);
-
-  const goToSummary = () => {
-    console.log('Summary', cashierDetails);
-    if (cashierDetails && cashierDetails.employerId) {
-      updateBankDetails(cashierBankDetails, cashierDetails.employerId);
-    } else {
-      toast.error('No cashier created yet.');
-    }
+  const goToBack = () => {
+    setCurrentComponent(ComponentState.BankDetails);
   };
 
   return (
-    <div className='w-full p-16 px-4 sm:px-6 lg:px-8'>
-      <div className='grid grid-cols-1 md:grid-cols-1 gap-6'>
-        {/* First Column */}
-        <div>
-          <label
-            htmlFor='bankName'
-            className='block text-sm font-medium text-black mt-4'
-          >
-            Bank Name
-          </label>
-          <input
-            type='text'
-            id='bankName'
-            className='mt-1 p-2 border-gray rounded-md w-full'
-            value={cashierBankDetails.bankName}
-            onChange={(e) =>
-              setCashierBankDetails({
-                ...cashierBankDetails,
-                bankName: e.target.value,
-              })
-            }
-          />
-
-          <label
-            htmlFor='branchName'
-            className='block text-sm font-medium text-black mt-4'
-          >
-            Branch Name
-          </label>
-          <input
-            type='text'
-            id='branchName'
-            className='mt-1 p-2 border-gray rounded-md w-full'
-            value={cashierBankDetails.bankBranchName}
-            onChange={(e) =>
-              setCashierBankDetails({
-                ...cashierBankDetails,
-                bankBranchName: e.target.value,
-              })
-            }
-          />
-
-          <label
-            htmlFor='accountNumber'
-            className='block text-sm font-medium text-black mt-4'
-          >
-            Account Number
-          </label>
-          <input
-            type='number'
-            id='accountNumber'
-            className='mt-1 p-2 border-gray rounded-md w-full'
-            value={cashierBankDetails.bankAccountNumber}
-            onChange={(e) =>
-              setCashierBankDetails({
-                ...cashierBankDetails,
-                bankAccountNumber: parseInt(e.target.value),
-              })
-            }
-          />
-
-          <label
-            htmlFor='additionalNotes'
-            className='block text-sm font-medium text-black mt-4'
-          >
-            Additional Notes
-          </label>
-          <textarea
-            id='additionalNotes'
-            className='mt-1 p-2 border-gray rounded-md w-full'
-            value={cashierBankDetails.employerDescription}
-            onChange={(e) =>
-              setCashierBankDetails({
-                ...cashierBankDetails,
-                employerDescription: e.target.value,
-              })
-            }
-          ></textarea>
-
-          <label
-            htmlFor='baseSalary'
-            className='block text-sm font-medium text-black mt-4'
-          >
-            Base Salary
-          </label>
-          <input
-            type='text'
-            id='baseSalary'
-            className='mt-1 p-2 border-gray rounded-md w-full'
-            value={cashierDetails.employerSalary}
-            readOnly
-          />
-        </div>
+    <div className='grid grid-cols-2 gap-4'>
+      <div className='bg-gray-100 p-4 rounded-lg'>
+        <p className='text-lg font-bold mb-2'>Personal Information</p>
+        <p>
+          <span className='font-semibold'>Name:</span>{' '}
+          {cashierDetails.employerFirstName} {cashierDetails.employerLastName}
+        </p>
+        <p>
+          <span className='font-semibold'>Nickname:</span>{' '}
+          {cashierDetails.employerNicName}
+        </p>
+        <p>
+          <span className='font-semibold'>Email:</span>{' '}
+          {cashierDetails.employerEmail}
+        </p>
+        <p>
+          <span className='font-semibold'>Phone Number:</span>{' '}
+          {cashierDetails.employerPhone}
+        </p>
+        <p>
+          <span className='font-semibold'>Address </span>{' '}
+          {cashierDetails.employerAddress}
+        </p>
+        <p>
+          <span className='font-semibold'>Date of Birth:</span>{' '}
+          {cashierDetails.dateOfBirth?.toString().slice(0, 10)}
+        </p>
       </div>
-      <div className='flex items-center justify-center gap-8 w-full'>
-        <button
-          type='button'
-          className={`text-white py-2.5 px-5 me-2 mb-2 rounded-lg ${
-            loading
-              ? 'bg-gray-500 cursor-not-allowed'
-              : 'bg-blueDarker hover:bg-blue'
-          }`}
-          onClick={goToSummary}
-          disabled={loading}
-        >
-          {loading ? 'Loading...' : 'Update'}
-        </button>
+      <div className='bg-gray-100 p-4 rounded-lg'>
+        <p className='text-lg font-bold mb-2'>Employment Details</p>
+        <p>
+          <span className='font-semibold'>Role:</span> {cashierDetails.role}
+        </p>
+        <p>
+          <span className='font-semibold'>Assign Branch:</span>{' '}
+          {cashierDetails.branchId}
+        </p>
+        <p>
+          <span className='font-semibold'>Base Salary:</span>{' '}
+          {cashierDetails.employerSalary}
+        </p>
+        <p>
+          <span className='font-semibold'>Pin:</span> {cashierDetails.pin}
+        </p>
+        <p>
+          <span className='font-semibold'>Bank Account Number:</span>{' '}
+          {cashierBankDetails.bankAccountNumber}
+        </p>
+        <p>
+          <span className='font-semibold'>Bank Name:</span>{' '}
+          {cashierBankDetails.bankName}
+        </p>
+        <p>
+          <span className='font-semibold'>Branch Name:</span>{' '}
+          {cashierBankDetails.bankBranchName}
+        </p>
+        <p>
+          <span className='font-semibold'>Additional Notes:</span>{' '}
+          {cashierBankDetails.employerDescription}
+        </p>
       </div>
+      <button
+        type='button'
+        className='py-2.5 px-5 me-2 mb-2 text-sm font-medium text-slate-900 focus:outline-none bg-white rounded-lg border border-gray hover:bg-gray'
+      >
+        <Link to='/manager-dashboard/Cashiers'>Done</Link>
+      </button>
+      <button
+        type='button'
+        className='text-white bg-blueDarker hover:bg-blue font-medium py-2.5 px-5 me-2 mb-2 rounded-lg'
+        onClick={goToBack}
+      >
+        Go To Bank Details
+      </button>
     </div>
   );
-};
+}
 
-export default CashierBankDetails;
+export default CashierDetailsSummary;
